third_party/WebKit/LayoutTests/external/wpt/html/webappapis/scripting/events/messageevent-constructor.https-expected.txt

This is a testharness.js-based test.
PASS Default event values
PASS MessageEventInit dictionary
PASS Passing null for ports member
PASS ports attribute should be a FrozenArray
PASS initMessageEvent operation
FAIL Passing null for ports parameter to initMessageEventassert_throws: function "function () {
    ev.initMessageEvent("test", true, false, "testData", "testOrigin", "testId", window, null)
  }" did not throw
FAIL initMessageEvent operation default parameter valuesassert_equals: MessageEvent.prototype.initMessageEvent.length should be 1 expected 1 but got 0
PASS Passing ServiceWorker for source member
Harness: the test ran to completion.
